+++ dabo/biz/dBizobj.py
@@ -141,17 +141,58 @@ class dBizobj(dObject):
 		return self._connection
 
 
-	def _flushUnchangedCursors(self):
-		"""Remove all cursors from this and all children, except current
-		and changed cursors."""
+	def clear(self, confirmed=False):
+		"""
+		Clear all cursors and records from self and children.
+
+		Use this when adding many records in batch, to keep memory use
+		from growing as records are added. For example:
+
+		for i in range(1000):
+			self.new(...)
+			for j in range(1000):
+				self.child.new(...)
+			self.save()
+			self.clear()
+
+		If you didn't issue clear() after each save(), you'd end up
+		with 1000 parent records, and 1000 cursors in the child pointing to
+		each of those 1000 parents, and each of those cursors would have
+		1000 rows in them.
+
+		Due to the destructive nature of this method, you must pass
+		confirmed=True. There will be no other warning.
+		"""
+		if not confirmed:
+			raise ValueError("Must pass call clearRecords(True) to "
+					"confirm that you are aware that data will be lost.")
+		self._flushCursors(flush_changed=True, flush_current=True)
+
+
+	def _flushCursors(self, flush_changed=False, flush_current=False):
+		"""
+		Remove cursors from this and all children.
+
+		By default, only unchanged non-current cursors are flushed.
+		"""
 		cursors = {}
 		for key, cursor in self.__cursors.items():
-			if cursor is self._CurrentCursor or cursor.isChanged():
+			if (not flush_current and cursor is self._CurrentCursor) \
+					or (not flush_changed and cursor.isChanged()):
 				cursors[key] = cursor
 		self.__cursors = cursors
-
+		if flush_current:
+			self.__currentCursorKey = None
 		for child in self._children:
-			child._flushUnchangedCursors()
+			child._flushCursors(flush_changed, flush_current)
+
+
+	def _flushUnchangedCursors(self):
+		"""
+		Remove all cursors from this and all children, except current
+		and changed cursors.
+		"""
+		self._flushCursors()
 
 
 	def getTempCursor(self, sql=None, params=None, requery=True):
@@ -1813,6 +1854,18 @@ class dBizobj(dObject):
 			child.Parent = self
 
 
+	def removeChild(self, child):
+		"""
+		Remove the passed child bizobj from this bizobj.
+
+		If the specified child bizobj isn't a child of this bizobj, raise
+		ValueError.
+		"""
+		children = self._children
+		child = children.pop(children.index(child))
+		child.Parent = None
+
+
 	def removeAllChildren(self):
 		"""
 		Remove all child bizobjs.
@@ -2420,6 +2473,10 @@ class dBizobj(dObject):
 	def addWhere(self, exp, comp="and"):
 		"""Add a filter expression to the where clause."""
 		return self._CurrentCursor.addWhere(exp, comp=comp)
+
+	def removeWhere(self, exp, comp="and"):
+		"""Remove a previously-added filter expression from the where clause."""
+		return self._CurrentCursor.removeWhere(exp, comp=comp)
 
 	def getSQL(self):
 		"""Returns the SQL statement currently set in the backend."""