+++ dabo/lib/datanav2/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-from dabo.lib.datanav import Form
-from dabo.lib.datanav import Grid
-from dabo.lib.datanav import Page, SelectPage, EditPage, BrowsePage, SortLabel
-from dabo.lib.datanav import IGNORE_STRING, SelectionOpDropdown
-from dabo.lib.datanav import PageFrameMixin, PageFrame
-from dabo.lib.datanav import Bizobj
-