+++ dabo/ui/uiwx/dGrid.py
@@ -635,7 +635,11 @@ class dColumn(dabo.ui.dPemMixinBase.dPemMixinBase):
 		Usually don't need this, but it helps to keep this in
 		line with other Dabo objects.
 		"""
-		self.Parent.removeColumn(self)
+		try:
+			self.Parent.removeColumn(self)
+		except ValueError:
+			# Will happen when the column has already been removed
+			pass
 
 
 	def _setAbsoluteFontZoom(self, newZoom):
@@ -2010,6 +2014,10 @@ class dGrid(cm.dControlMixin, wx.grid.Grid):
 #		self.bindEvent(dEvents.GridContextMenu, self._onContextMenu)
 		self.bindEvent(dEvents.GridMouseRightClick, self._onGridMouseRightClick)
 		self.bindEvent(dEvents.Resize, self._onGridResize)
+
+		self.bindEvent(dEvents.Create, self._onCreate)
+		self.bindEvent(dEvents.Destroy, self._onDestroy)
+
 		super(dGrid, self)._initEvents()
 
 
@@ -2045,7 +2053,7 @@ class dGrid(cm.dControlMixin, wx.grid.Grid):
 		Call this when your datasource or dataset has changed to get the grid showing
 		the proper number of rows with current data.
 		"""
-		# Note that we never call self.super(), because we don't need/want that behavior.
+		# We never call the superclass update, because we don't need/want that behavior.
 		last = getattr(self, "_lastCellSelectedTime", 0)
 		cur = time.time()
 		if cur - last < .5:
@@ -2296,6 +2304,8 @@ class dGrid(cm.dControlMixin, wx.grid.Grid):
 
 
 	def _restoreSort(self):
+		if not self.Sortable:
+			return
 		self.sortedColumn = self._getUserSetting("sortedColumn")
 		self.sortOrder = self._getUserSetting("sortOrder")
 
@@ -2305,7 +2315,7 @@ class dGrid(cm.dControlMixin, wx.grid.Grid):
 				if col.DataField == self.sortedColumn:
 					sortCol = idx
 					break
-			if sortCol is not None:
+			if sortCol is not None and col.Sortable:
 				if self.RowCount > 0:
 					self.processSort(sortCol, toggleSort=False)
 
@@ -2465,6 +2475,11 @@ class dGrid(cm.dControlMixin, wx.grid.Grid):
 			self.autoSizeCol("all", True)
 		return True
 
+	def _onCreate(self, evt):
+		self.restoreDataSet()
+
+	def _onDestroy(self, evt):
+		self.saveDataSet()
 
 	def _onGridResize(self, evt):
 		# Prevent unnecessary event processing.
@@ -3039,6 +3054,20 @@ class dGrid(cm.dControlMixin, wx.grid.Grid):
 		dabo.ui.callAfterInterval(200, self.Form.update)  ## rownum in status bar
 
 
+	def restoreDataSet(self):
+		if self.SaveRestoreDataSet:
+			ds = self.Application.getUserSetting("%s.DataSet"
+					% self.getAbsoluteName())
+			if ds is not None:
+				self.DataSet = ds
+
+
+	def saveDataSet(self):
+		if self.SaveRestoreDataSet:
+			self.Application.setUserSetting("%s.DataSet"
+					% self.getAbsoluteName(), self.DataSet)
+
+
 	def runIncSearch(self):
 		"""Run the incremental search."""
 		gridCol = self.CurrentColumn
@@ -3365,7 +3394,7 @@ class dGrid(cm.dControlMixin, wx.grid.Grid):
 			columns = columns[0]
 		return [self._resolveColumn(col) for col in columns]
 
-		
+
 	def _resolveColumn(self, colOrIdx, returnColumn=True, logOnly=False):
 		"""
 		Accepts either a column object or a column index, and returns a column
@@ -4281,15 +4310,8 @@ class dGrid(cm.dControlMixin, wx.grid.Grid):
 
 
 	def __onWxGridRowSize(self, evt):
-		if not self.ResizableRows:
-			evt.Veto()
-			ht = self._rowHeight
-			# This will force a refresh
-			self._rowHeight = -1
-			self.RowHeight = ht
-		else:
-			self.raiseEvent(dEvents.GridRowSize, evt)
-			evt.Skip()
+		self.raiseEvent(dEvents.GridRowSize, evt)
+		evt.Skip()
 
 
 	def __onWxHeaderContextMenu(self, evt):
@@ -4366,7 +4388,6 @@ class dGrid(cm.dControlMixin, wx.grid.Grid):
 
 
 	def __onWxHeaderPaint(self, evt):
-		self.raiseEvent(dEvents.GridHeaderPaint, evt)
 		updateBox = self._getWxHeader().GetUpdateRegion().GetBox()
 		self._paintHeader(updateBox)
 
@@ -4774,7 +4795,14 @@ class dGrid(cm.dControlMixin, wx.grid.Grid):
 		return self._resizableRows
 
 	def _setResizableRows(self, val):
-		self._resizableRows = val
+		if self._constructed():
+			self._resizableRows = val
+			if val:
+				self.EnableDragRowSize()
+			else:
+				self.DisableDragRowSize()
+		else:
+			self._properties["ResizableRows"] = val
 
 
 	def _getRowColorEven(self):
@@ -4853,6 +4881,13 @@ class dGrid(cm.dControlMixin, wx.grid.Grid):
 
 	def _setSameSizeRows(self, val):
 		self._sameSizeRows = bool(val)
+
+
+	def _getSaveRestoreDataSet(self):
+		return getattr(self, "_saveRestoreDataSet", False)
+
+	def _setSaveRestoreDataSet(self, val):
+		self._saveRestoreDataSet = bool(val)
 
 
 	def _getSearchable(self):
@@ -5265,6 +5300,17 @@ class dGrid(cm.dControlMixin, wx.grid.Grid):
 	SameSizeRows = property(_getSameSizeRows, _setSameSizeRows, None,
 			_("""Is every row the same height?  (bool)"""))
 
+	SaveRestoreDataSet = property(_getSaveRestoreDataSet, _setSaveRestoreDataSet, None,
+			_("""Specifies whether the DataSet is persisted to preferences (bool).
+
+				This allows you to build a grid to capture user input of some form, and
+				instead of saving the row and field values to a database, to save the
+				entire dataset to a single key in the prefs table.
+
+				Use this sparingly for grids that won't grow too large.
+
+				The default is False."""))
+
 	Searchable = property(_getSearchable, _setSearchable, None,
 			_("""Specifies whether the columns can be searched.   (bool)
 
@@ -5421,7 +5467,6 @@ class _dGrid_test(dGrid):
 		self.Editable = False
 		#self.Sortable = False
 		#self.Searchable = False
-
 
 	def afterInit(self):
 		super(_dGrid_test, self).afterInit()